Get jest to use it's tsconfig again
