+++ jest.config.js
@@ -16,6 +16,7 @@ module.exports = {
       // ts-jest configuration goes here
       diagnostics: false,
       isolatedModules: true,
+      tsconfig: "./test/tsconfig.jest.json",
     },
   },
 };
