+++ test/tsconfig.jest.json
@@ -1,0 +1,9 @@
+
+{
+  "compilerOptions": {
+    "strictNullChecks": false,
+    "noImplicitAny": false,
+  },
+  "extends": "../tsconfig.json",
+  "include": ["../src", "../test"]
+}
