--- tsconfig.test.json
@@ -1,28 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2015",
-    "allowJs": true,
-    "jsx": "react",
-    "module": "commonjs",
-    "skipLibCheck": true,
-    "esModuleInterop": true,
-    "noImplicitAny": true,
-    "sourceMap": true,
-    "baseUrl": ".",
-    "outDir": "dist",
-    "moduleResolution": "node",
-    "resolveJsonModule": true,
-    "strictNullChecks": true,
-    "strict": true,
-    "paths": {
-      "*": [
-        "node_modules/*"
-      ]
-    },
-    "rootDir": "."
-  },
-  "include": [
-    "src",
-    "test"
-  ],
-}
